fix(structure): move teasers to "helpers"
