+++ src/components/helpers/initGalleryTeasers.js
