+++ src/components/helpers/initVideoTeasers.js
